requirements.txt

PyQt5
google-generativeai
